Moved Ninchat activity to ninchatactivity package
