+++ ninchatsdk/src/androidTest/java/com/ninchat/sdk/espresso/ninchatqueuelist/view/NinchatOpenListAdapter.kt
@@ -3,18 +3,14 @@ package com.ninchat.sdk.espresso.ninchatqueuelist.view
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 import androidx.test.core.app.ActivityScenario
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.BoundedMatcher
-import androidx.test.espresso.matcher.ViewMatchers.withId
-import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
 import androidx.test.platform.app.InstrumentationRegistry
 import com.ninchat.sdk.NinchatSession
 import com.ninchat.sdk.NinchatSessionManager
 import com.ninchat.sdk.R
-import com.ninchat.sdk.activities.NinchatActivity
+import com.ninchat.sdk.ninchatactivity.view.NinchatActivity
 import com.ninchat.sdk.ninchatqueuelist.model.NinchatQueue
 import org.hamcrest.Description
 import org.hamcrest.Matcher
