+++ ninchatsdk/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
 
     <application>
         <activity
-            android:name=".activities.NinchatActivity"
+            android:name=".ninchatactivity.view.NinchatActivity"
             android:theme="@style/NinchatTheme" />
         <activity
             android:name=".ninchatqueue.view.NinchatQueueActivity"
